backend/requirements.txt
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.1
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pytest==7.3.2
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
